pyproject.toml
[project]
name = "tgmusicbot"
version = "1.2.1"
description = "Telegram Music Bot with full playback support using PyTdBot and PyTgCalls"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Ashok Shau", email = "ashokshau@t.me" },
]
keywords = [
    "telegram",
    "music",
    "bot",
    "async",
    "tdlib",
    "pytgcalls",
    "voicechat",
    "spotify",
    "ytmusic",
    "opensource",
    "pytdbot",
    "py-tgcalls",
    "yt-dlp",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Chat :: Telegram",
    "Topic :: Multimedia :: Sound/Audio :: Players",
    "Environment :: Console",
    "Framework :: AsyncIO",
]
dependencies = [
    "aiofiles~=24.1.0",
    "APScheduler~=3.11.0",
    "cachetools~=5.3.2", # Diperbarui dari "~=6.0.0" ke "~=5.3.2"
    "kurigram~=2.2.4",
    "meval~=2.5",
    "ntgcalls~=2.0.2",
    "pillow~=11.2.1",
    "psutil~=7.0.0",
    "py-yt-search~=0.3",
    "pycryptodome~=3.23.0",
    "pydantic~=2.11.5",
    "pymongo~=4.13.1",
    "py-tgcalls~=2.2.2",
    "pytgcrypto~=1.2.11",
    "python-dotenv~=1.1.0",
    "pytdbot~=0.9.4",
    "pytz~=2025.2",
    "tdjson~=1.8.50",
    "ujson~=5.10.0",
    "yt-dlp~=2025.6.9",
    "google-generativeai",
    "requests",
]

[project.license]
file = "LICENSE"

[project.optional-dependencies]
dev = [
    "setuptools~=80.9.0",
    "black",
    "ruff",
]

[project.scripts]
tgmusic = "src.__main__:main"

[project.urls]
Homepage = "https://github.com/AshokShau/tgmusicbot"
Source = "https://github.com/AshokShau/tgmusicbot"
"Issue Tracker" = "https://github.com/AshokShau/tgmusicbot/issues"

[tool.uv]
package = true

[tool.setuptools]
packages = [
    "src",
]
